# Constant propogation